Code complete, errors persist
Well, the code is complete for adding inventory to each scene, however I continue to have the error where nothing is interactable. Will try to get that working once Elinda has added/modified the ray cast code so that we're casting from the mouse position instead of the camera position.